Fix .jshintrc location logic
The code to look for .jshintrc intended to start from the current file's
directory and walk up on the filesystem until a .jshintrc was found.
Unfortunately this code was only ever looking at the current directory
and not really recursing up. Fix is to actually use the path passed to
FindRC().